Display: Implement partial rendering
Only for the time shown so far, as this is the most important update.

Note: When extending this to all cases, you'd somehow have to allow for multiple updates in one render. Easiest solution is to just set DisplayUpdate::all if _needUpdate is already something else than none and do a full render then.